+++ aws/driver/gen_drivers.go
@@ -593,6 +593,18 @@ func (d *EcsDriver) Lookup(lookups ...string) (driverFn driver.DriverFn, err err
 		}
 		return d.Delete_Containercluster, nil
 
+	case "startcontainerservice":
+		if d.dryRun {
+			return d.Start_Containerservice_DryRun, nil
+		}
+		return d.Start_Containerservice, nil
+
+	case "createcontainer":
+		if d.dryRun {
+			return d.Create_Container_DryRun, nil
+		}
+		return d.Create_Container, nil
+
 	default:
 		return nil, driver.ErrDriverFnNotFound
 	}