+++ aws/gen_api.go
@@ -105,6 +105,8 @@ var ResourceTypes = []string{
 	"scalingpolicy",
 	"repository",
 	"containercluster",
+	"containerservice",
+	"container",
 	"user",
 	"group",
 	"role",
@@ -167,6 +169,8 @@ var ServicePerResourceType = map[string]string{
 	"scalingpolicy":       "infra",
 	"repository":          "infra",
 	"containercluster":    "infra",
+	"containerservice":    "infra",
+	"container":           "infra",
 	"user":                "access",
 	"group":               "access",
 	"role":                "access",
@@ -210,6 +214,8 @@ var APIPerResourceType = map[string]string{
 	"scalingpolicy":       "autoscaling",
 	"repository":          "ecr",
 	"containercluster":    "ecs",
+	"containerservice":    "ecs",
+	"container":           "ecs",
 	"user":                "iam",
 	"group":               "iam",
 	"role":                "iam",
@@ -297,6 +303,8 @@ func (s *Infra) ResourceTypes() []string {
 		"scalingpolicy",
 		"repository",
 		"containercluster",
+		"containerservice",
+		"container",
 	}
 }
 
@@ -333,6 +341,8 @@ func (s *Infra) FetchResources() (*graph.Graph, error) {
 	var scalingpolicyList []*autoscaling.ScalingPolicy
 	var repositoryList []*ecr.Repository
 	var containerclusterList []*ecs.Cluster
+	var containerserviceList []*ecs.TaskDefinition
+	var containerList []*ecs.Container
 
 	errc := make(chan error)
 	var wg sync.WaitGroup
@@ -705,6 +715,38 @@ func (s *Infra) FetchResources() (*graph.Graph, error) {
 	} else {
 		s.log.Verbose("sync: *disabled* for resource infra[containercluster]")
 	}
+	if s.config.getBool("aws.infra.containerservice.sync", true) {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			var resGraph *graph.Graph
+			var err error
+			resGraph, containerserviceList, err = s.fetch_all_containerservice_graph()
+			if err != nil {
+				errc <- err
+				return
+			}
+			g.AddGraph(resGraph)
+		}()
+	} else {
+		s.log.Verbose("sync: *disabled* for resource infra[containerservice]")
+	}
+	if s.config.getBool("aws.infra.container.sync", true) {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			var resGraph *graph.Graph
+			var err error
+			resGraph, containerList, err = s.fetch_all_container_graph()
+			if err != nil {
+				errc <- err
+				return
+			}
+			g.AddGraph(resGraph)
+		}()
+	} else {
+		s.log.Verbose("sync: *disabled* for resource infra[container]")
+	}
 
 	go func() {
 		wg.Wait()
@@ -1073,6 +1115,36 @@ func (s *Infra) FetchResources() (*graph.Graph, error) {
 			}
 		}()
 	}
+	if s.config.getBool("aws.infra.containerservice.sync", true) {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			for _, r := range containerserviceList {
+				for _, fn := range addParentsFns["containerservice"] {
+					err := fn(g, r)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}
+			}
+		}()
+	}
+	if s.config.getBool("aws.infra.container.sync", true) {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			for _, r := range containerList {
+				for _, fn := range addParentsFns["container"] {
+					err := fn(g, r)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}
+			}
+		}()
+	}
 
 	go func() {
 		wg.Wait()
@@ -1158,6 +1230,12 @@ func (s *Infra) FetchByType(t string) (*graph.Graph, error) {
 		return graph, err
 	case "containercluster":
 		graph, _, err := s.fetch_all_containercluster_graph()
+		return graph, err
+	case "containerservice":
+		graph, _, err := s.fetch_all_containerservice_graph()
+		return graph, err
+	case "container":
+		graph, _, err := s.fetch_all_container_graph()
 		return graph, err
 	default:
 		return nil, fmt.Errorf("aws infra: unsupported fetch for type %s", t)