+++ graph/resourcetest/resourcetest.go
@@ -145,6 +145,14 @@ func ContainerCluster(id string) *rBuilder {
 	return new("containercluster", id).Prop(properties.ID, id)
 }
 
+func ContainerService(id string) *rBuilder {
+	return new("containerservice", id).Prop(properties.ID, id)
+}
+
+func Container(id string) *rBuilder {
+	return new("container", id).Prop(properties.ID, id)
+}
+
 func (b *rBuilder) Prop(key string, value interface{}) *rBuilder {
 	b.props[key] = value
 	return b